Run this Script with Unclipped Data
This is the most current verison of the master script WITHOUT Matt's DEM portion.  This code clips each input layer by the boundaries, runs specific euclidean and reclassify functions.  It then converts the rasters to tif format and converts them to a numpy array.  The numpy arrays are appended into one list, and are attempted to be added all together (I believe it fails at this).  It then attempts to write this added numpy array to a TIF file using rasterio.  The ouput is a strange raster with values ranging from 1-22.  The correct final raster should only range from 0-to the nunber of inputs given.