Suitability Raster Created
So I got the code to run with all the inputs but the values of the value raster are all screwed up.  The values range from 0-121.  They should only range from 0-4.  I think something is happening with the reclassify functions.  We may have to reclassify using np.where once the rasters are in array format.  We will look into this during class tonight.