Script with clipping and without clipping
The without clipping can be ran with already pre-clipped files after you run the clipping portion of script_with.  We need to find a way to add an extra row of 1's to the slope array.  While the reclassify tool does get the shape right for the other 3 arrays, it is having unintended consequences and returning the wrong values.  Long story short, we can't rely on the Reclassify tool.